[SxS] Handle ungrouping tab in split view via extensions
A split view has to have both tabs in a group or not in a group.
Currently if a user removes a single tab from a group it causes a crash.
There are two ways to solve this:

1. Remove both tabs within the split from the group
2. Unsplit the tabs and just remove one from the group

I've went with the first option because it avoids destroying the split
and to the user the split view appears as a single tab. I'm open to
going with option 2 if that would be preferred from an extensions
perspective.

Bug: 458267485
Change-Id: I74c103653750ee915c00d23b0b1e4b522e00dc45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7126988
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Commit-Queue: Alison Gale <agale@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1541805}