[iOS][SyncedSetUp] Add Impression Limit
This change introduces a new integer profile pref to count
impressions for the Synced Set Up Flow. A Finch parameter is added for
the `kIOSSyncedSetUp` feature that introduces an impression limit for
the UI. The Synced Set Up UI will not be shown to the user after
reaching the impression limit.

R=bwwilliams@google.com

Fixed: 458415548
Change-Id: I0484369ff3fec80d89c9146fe8e956d991ea876f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7127644
Reviewed-by: Benjamin Williams <bwwilliams@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Eric Ekey <ericekey@google.com>
Cr-Commit-Position: refs/heads/main@{#1541822}