[Tab Model] Expose onDestroy method on TabModelObserver
In order to fix the packager cache's lifecycle we need to observe when
the TabModel is destroyed.

Also ensure tab closures are committed on destroy in
TabCollectionTabModelImpl. In theory, this is a non-issue as both
pausing and a clean shutdown invoke this upstack, but for consistency
with the TabModelImpl implementation there is no harm in also doing it
here.

Bug: 458679374
Change-Id: Ie14863e90acb4208be3c70c4d2e7fb807aa41a5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7131818
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Fiaz Muhammad <mfiaz@google.com>
Cr-Commit-Position: refs/heads/main@{#1541838}