[spc][bbk] Allow the SPC service to use BBKs on Desktop
Note that the BBK is still hidden behind the
`blink::features::kSecurePaymentConfirmationBrowserBoundKeys` feature
flag.

Changed PasskeyBrowserBinder::CreateUnboundKey to use the BBK ID of
the BBK returned from the key store rather than the BBK ID provided.
This is important for desktop because new keys ignore the provided
ID and generate one on creation of the BBK.

Changed SecurePaymentConfirmationService::OnAuthenticatorMakeCredential
to provide a last_used value of the current system time when creating
a new BBK. Note that this is only needed for Windows hence the buildflag.

Bug: 441552524
Change-Id: I7da1adb4b781323aab6fc5d9935a9811640d9024
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7125163
Commit-Queue: Darwin Yang <darwinyang@chromium.org>
Reviewed-by: Slobodan Pejic <slobodan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1541848}