Remove redundant map lookups in prefetch_match_resolver.cc.
Instead of doing a `candidates_.contains(key)` followed by
`candidates_[key]`, reuse the iterator from the initial container check.
This avoids the duplicate lookup and need to re-calculate the hash from
the strings in the key, and also avoids default-constructing the default
value when using the bracket operators.

Also, change `candidate_data` to be a pointer instead of a reference to
a `std::unique_ptr` so that it doesn't need to dereference twice to
access methods in the pointer.

Bug: 452950815
Change-Id: I72847a9bd698167f5f69b8ec3a952fc1060574d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7088031
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Jeffrey Yu <yuje@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1541859}