Skip the pre-launch dialog if the disambiguation dialog was shown
This CL suppresses the secondary protocol handler dialog for an app
if the user was previously routed through the main dialog; the decision
is also persisted into the web app system if the dialog was accepted
with the "remember selection" checkbox.

Note that chrome://settings/handlers can still be used to reset the
approval for a particular IWA, and it might be possible to artificially
run into a scenario where only the secondary dialog is shown.

Bug: 458271211
Change-Id: I9986cd2e16454bc513f0dce862b631b40e8b6ff9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7123790
Reviewed-by: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Andrew Rayskiy <greengrape@google.com>
Cr-Commit-Position: refs/heads/main@{#1541868}