Add Structured-test-id experiment to chromiumos builder group
The experiment changes how results are displayed and processed by
resultdb. Failures should be deterministic with a problem or failure to upload to resultdb.

Change was reverted in https://chromium-review.googlesource.com/c/chromium/src/+/7097758 because of a failure in chrome_sizes. That failure was fixed in:

https://chromium-review.googlesource.com/c/chromium/src/+/7097590

Not doing a reland to separate the android and chromeos builder groups since android had a revert for a separate reason earlier.

Bug: 454427565
Change-Id: I0539c939222db8e1b4dda779ca653f88efd625ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7128672
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Benjamin Joyce (Ben) <bjoyce@google.com>
Cr-Commit-Position: refs/heads/main@{#1541883}