[Media Controls] Use saturated_cast to safely format media time
For the Blink media controls, we currently cast the length of the media
from a double to an int. This is undefined behavior whenever the double
falls outside of the range of an int. This CL changes the static_cast
to a saturated_cast to avoid the undefined behavior.

Bug: 41492424
Change-Id: I279292c6db3600b71942fa0f13466b218ff26b2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7129698
Auto-Submit: Tommy Steimel <steimel@chromium.org>
Commit-Queue: Frank Liberato <liberato@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1541888}