Add two new interfaces on IElevator
This CL creates a new IElevator2 which adds two new interfaces. The
first is for launching a new isolated browser and the second is for
using mojo over the COM interface. A longer term plan is to move the
existing COM interfaces to mojo so this provides that interface so
hopefully the COM interface never has to change again.

Both are left unimplemented for now.

To support version skew - the new IElevator2 has a different IID but
also supports the old IElevator interface: Newer browser versions using
this new IID will connect to the IElevator2 but earlier versions of
browser before this version will connect to IElevator. This makes the
code compatible with old versions of browser connecting to new versions
of the elevation service.

New versions of the browser needing IElevator2 connecting to an older
version of the elevation service will fail to obtain the COM interface
since IElevator2 is not implemented - this should not happen in theory
since both are updated at the same time.

The app-bound browser tests are augmented to use both the old
(IElevator) and new (IElevator2) IIDs to verify that both work from the
same elevation service, however the production code is updated to always
attempt to use IElevator2.

BUG=383157189,433545123

Change-Id: I6975d994a3b7c317f7e22edb81182e51158fb36f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6813894
Reviewed-by: S Ganesh <ganesh@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1541900}