WebXr Layers API: Add is_static support for layers.
Create a single image swapchain for static layers. Do not propagate the
shared buffer after the first drawing for a static layer. Do not show
the layer until JavaScript requests a drawing buffer for the layer.
Propagate the needsRedraw state for a newly created layer and for a
newly added layer to the render state.

Bug: 454325950
Change-Id: Ib2e2872e0f5f77690a1ecc9802cb3a970f7cc527
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7121918
Commit-Queue: Evgenii Alekseev (xWF) <ealekseev@google.com>
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1541908}