[NTP Theme] Add metrics for pinch in preview dialog.
This CL adds a bucket "PinchToResize" to metric
"NewTabPage.Customization.Theme.UploadImage.PreviewInteractions" to
record when user scales or scrolls to resize the image in the preview
dialog. We only record it once regardless of how many times user scales
or scrolls the image, or whether user clicks the cancel or save button.

This histogram is recorded as
https://drive.google.com/file/d/1VHkaqUvvsQrsOyvOHo7CJi7FKBCwgHK1/view?usp=sharing

Bug: 456173872
Change-Id: I13bf1ca4077f23ceace728c465d8fa9105a230d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7128599
Reviewed-by: Yanling Hu <yanlinghu@google.com>
Reviewed-by: Xi Han <hanxi@chromium.org>
Commit-Queue: Yanting Yu <yyanting@google.com>
Cr-Commit-Position: refs/heads/main@{#1541917}