PaintTiming: Fix not setting largest-painted-image for first video frame
With ReportFirstFrameTimeAsRenderTime enabled, NotifyFirstVideoFrame()
sets the paint time directly, rather than setting it in
AssignPaintTimeToRegisteredQueuedRecords(). But there's an early-out
when assigning paint times if the paint time is already set, which
prevents us from both setting the largest_painted_image_ to the
video ImageRecord and removing the record from the pending set. This
CL removes that condition, as well as a couple other that are obsolete.

This only affects ReportFirstFrameTimeAsRenderTime, which has not yet
shipped. Also, aside from the memory leak, it doesn't affect LCP since
largest_pending_image_ is set --- but this is a latent issue, which I
discovered while working towards emitting more intermediate candidates.

Bug: 454067883
Change-Id: I5d09529a76abb5aa4ce9333c08e8ef7ebdbb029f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7128129
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1541941}