Make a MemoryPressureListener for MemoryReclaimer
The goal of this change is to separate all the memory pressure
interventions in RenderThreadImpl::OnMemoryPressure so that they can
be invoked separately.

This is a necessary change for migrating all MemoryPressureListeners
to the base::MemoryConsumer interface, where each intervention is
tagged with different properties.

Bug: 436324601
Change-Id: Ie6c3fe4ae2f6ce36612a913ef181d5d478fef1d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7127962
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1541975}