[Chrome Next] Convert main container to shadow overlay
This changes the main container to be a separate, floating, translucent
shadow overlay, with no children.

This is done in order to flatten the view hierarchy, which is necessary
to be able to split the top container between inside and outside the
"main" area when the toolbar-height side panel is visible.

It applies the absolute fewest code changes possible to the layouts to
get the feature working without any regressions.

Follow-up changes will:
 - Tidy up the layout to reflect the new view hierarchy.

Bug: 453717426
Change-Id: I48c9e3117fe78019b254bd95d2f61ec723660d6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7124998
Reviewed-by: Darryl James <dljames@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1541994}