Pass contents_view to `ClientViewFactory` callback during widget initialization
By passing contents_view to callback, it makes it easier for clients to
create Custom ClientView.

With the previous interface, it was quite difficult to use
`WidgetDelegate::SetContentsView()` along with `ClientViewFactory`. A
user would have to either bind the delegate to the `ClientViewFactory`
in order to get the contents via `TransferOwnershipOfContentsView()` or
a user would not use `WidgetDelegate::SetContentsView()` and
initialize/bind the contents_view directly to the callback. Both of
these patterns are makes it difficult to use the `ClientViewFactory`
callback.

Bug: b:458452855
Change-Id: Ice76e82bee56810d93e4efe34c5fd4dfad8b4d9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7128133
Commit-Queue: Zoraiz Naeem <zoraiznaeem@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542026}