[glic] Simplify widget setup in `GlicWidget::Create()`
In essence, when widget type is
`views::Widget::InitParams::TYPE_WINDOW`, we create a NonClientView,
which in turn creates a FrameView and ClientView, whereas for a
`views::Widget::InitParams::TYPE_WINDOW_FRAMELESS` widget only has
contents_view.

Based on this underlying fact, this CL renames UseClientView() to
CreateNonClientView(),

Also, instead of specifying content_view after widget creation, it is
done via WidgetDelegate interface.

Additionally, we do not a custom ClientView with a its awkward
initialization in GlicWidgetDelegate.

In a future CL, we will fix the lifetime management of WidgetDelegate
and get rid of `GlicWidgetDelegate`

Bug: b:458115863
Change-Id: I38371c87d4c75ead359714b508debacb08da11a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7125107
Reviewed-by: Ian Wells <iwells@chromium.org>
Commit-Queue: Zoraiz Naeem <zoraiznaeem@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542045}