[Extensions] Remove incognito check when showing install dialog
Checking for incognito profile when finding the browser to show the
extension install dialog was added a long time ago as a workaround for
an escape hatch. Since then, extensions cannot be directly installed
on incognito. Therefore, we can remove this workaround and CHECK that
a browser is returned.

Bug: 40319332
Change-Id: I8bcce45de1403e4d28039f1fdd3e218ebe8441e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7126990
Reviewed-by: Tim <tjudkins@chromium.org>
Auto-Submit: Emilia Paz <emiliapaz@chromium.org>
Commit-Queue: Emilia Paz <emiliapaz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542075}