Remove the location bar attachments group.
The group is no longer serving its purpose in compact ui, exposing
the button that we want to remain hidden. The button visibility
continues to be maintained by the updateModeSelectorVisibility, but
the group no longer helps, occasionally exposing the button we want
to remain hidden.

Change-Id: I0a1d817a5332dccbfb6b5c8dfdebbf1883a75417
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7132634
Auto-Submit: Tomasz Wiszkowski <ender@google.com>
Reviewed-by: Patrick Noland <pnoland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542089}