[Robust Window Management] Allows for multiple inactive windows to be selected.
This CL is the pre-work necessary to allow for multiple windows to be closed at once via the window manager. If the robust window management flag is enabled, a user will be able to select more than one window. If `mSelectedItems` != 1, then the restore button is disabled. This capability is not available for active instances and that behavior remains unchanged.

Bug: 456472469
Verification: http://screencast/cast/NDgwMjQ1NjMxMDUxMzY2NHxhNmE0ODYxNi0zNQ
Change-Id: Ia454590e4a0f20432499e0697d2232d86825fa47
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7120368
Reviewed-by: Aishwarya Rajesh <aishwaryarj@google.com>
Commit-Queue: Ryan Thomas <rathomas@google.com>
Cr-Commit-Position: refs/heads/main@{#1542093}