[flex] Avoid FreezeViolations in LineFlexer constructor.
FreezeViolations is currently an issue as floats aren't associative.
https://en.wikipedia.org/wiki/Associative_property#Nonassociativity_of_floating-point_calculation

Currently we'll add floats together, then subtract them causing issues.

For the loop in the constructor we don't need to invoke FreezeViolations
at all, we can simply freeze the items, assign the flexed_content_size,
and skip the rest of the logic in the loop.

Additionally remove passing the sum_flex_base_size to the line-flexer we
can just calculate the remaining free-space again.

Change-Id: Ife6fc30b50712ee5bfa42beeb6828f3ef4840aea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7131613
Reviewed-by: David Grogan <dgrogan@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542120}