Replace old PipeWire CIPD package with the new one
We introduced new PipeWire CIPD package as pipewire-1.4 as a mid-step to
allow migrate existing tests in WebRTC. The tests have been now migrated
to use pipewire-1.4, but the goal is to have just one CIPD package with
the latest version. Since the tests have been migrated, the old CIPD
package is unused and can be now updated to the latest version, same as
we have in pipewire-1.4. The next step will be to update the tests in
WebRTC to back use pipewire CIPD package and as a follow-up we can get
rid of the pipewire-1.4 CIPD package.

Also remove PipeWire Media Session that has been replaced by Wireplumber

Bug: 414491370
Change-Id: I9111e3d0c35c51bc8a0a60f391ae47f151f73e55
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7130859
Reviewed-by: Mirko Bonadei <mbonadei@chromium.org>
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Jan Grulich <grulja@gmail.com>
Cr-Commit-Position: refs/heads/main@{#1542156}