Make documentURI non-nullable to match DOM spec
This patch updates the documentURI attribute from nullable to
non-nullable to match the DOM specification.
In [1], the URL was modified to always return a non-nullable value.
Therefore, since the current implementation already satisfies the
non-nullable requirement, this patch is not an actual behavior change.

[1] https://codereview.chromium.org/2768373005

Bug: 456874035
Change-Id: I54bf9341cfa8c947cb77d8243d39a0002e164295
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7110200
Reviewed-by: Jinho Bang <zino@chromium.org>
Commit-Queue: YeongHan Kim <soosungp33@gmail.com>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542185}