Introduce NavigatorPlugins mixin
This patch migrates a partial interface to an interface mixin
to aligin with the specification.

Originally, this partial interface was introduced as an workaround
for a layering violation between core/ and modules/, since Blink
could not include mixins across those layers.

Now that the IDL compiler supports cross-component mixins[1], the
workaround is no longer needed, and this patch replaces the partial
interface with interface mixin to follow the specs[2].

This CL has no behavior changes.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/6968299
[2] https://html.spec.whatwg.org/multipage/system-state.html#navigatorplugins

Bug: 453570181
Change-Id: If86b1d5ea3acd15300501dff4f114df303ad5f0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7123899
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Jinho Bang <zino@chromium.org>
Commit-Queue: Suyeon Ji <zeesuyeon@gmail.com>
Cr-Commit-Position: refs/heads/main@{#1542267}