Add update_manifest_url to app manifest
This CL adds an update_manifest_url_field as described in the
design doc. This field is used in isolated web apps and points
to the update manifest that is externally hosted and provides
a list of available updates for a given app. This is required
for IWA user installations to make sure that the app can be kept
up to date.

DD: go/unmanaged-iwa
Bug: 457651205
Change-Id: I35bbac5443eae5dadb209eef13d8ebff0b54e57a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7117860
Reviewed-by: Magdalena Skarbińska <mskarbinska@google.com>
Commit-Queue: Simon Hangl <simonha@google.com>
Reviewed-by: Antonio Sartori <antoniosartori@chromium.org>
Reviewed-by: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542392}