[Omnibox AIM] Add UnusedKeyboardEventHandler to omnibox popup
Command events like selectAll or copy/paste were not being handled in
the AIM popup. This is because when keyboard events are not handled in
the renderer, it needs to be redispatched in the browser. This is
handled by the UnhandledKeyboardEventHandler that was missing in the
omnibox popup.

Bug: 459236284
Change-Id: Iaabcf9b610c3c7548e4f0d18765827da90f9d153
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7136498
Auto-Submit: David Yeung <dayeung@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542397}