[iOS][CX] Add passkey importer
Introduces a class responsible for handling passkey import process. The
logic is currently only needed on //ios, but it might eventually be
reused, so it's implemented in //components straightaway.

High-level logic of the importer implenented in this CL:
1) receives vector<WebuthnCredentialsSpecifics> as input
2) asynchronously parses each passkey and assigns it to one of the
(invalid, conflicting, valid) groups
3) informs the caller about invalid and conflicting passkeys by passing
necessary display info to present in the UI
4) caches valid and conflicting passkeys and waits for the caller to
finalise the import process

Conflict is defined as an incoming passkey having matching (rp_id,
user_id) pair. The user will be able to choose in the UI whether they
want to "override" existing passkey with an incoming one (i.e. existing
passkey will become a shadow of the incoming one).

Validating incoming passkey fields will be implemented in a follow-up.

The follow-up CL will add an API for the caller to resume/finalise the
import by passing the info about which of the conflicting passkey pairs
to choose.

Bug: 458337350
Change-Id: I19a062c05a92733614f528442d3670d144e6fad2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7123074
Commit-Queue: Rafał Godlewski <rgod@google.com>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/main@{#1542446}